backend/enrich_data.py
import requests
from bs4 import BeautifulSoup
import sqlite3
import os
import time

# Note: In a real IEEE research study, we minimize external dependencies
# the easiest way to get "actual movie details" without an API key
# is to scrape a movie search portal or use a public movie DB.
# This script enriches our local DB so that energy tests remain controlled.

DB_PATH = os.path.join(os.path.dirname(__file__), 'database.db')

def get_poster_from_tmdb(movie_title, year):
    search_query = f"{movie_title} {year}"
    # TMDB search URL
    url = f"https://www.themoviedb.org/search?query={requests.utils.quote(search_query)}"
    headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
    
    try:
        response = requests.get(url, headers=headers)
        if response.status_code != 200:
            return None
        
        soup = BeautifulSoup(response.text, 'html.parser')
        # Find first result poster
        image = soup.find('img', class_='poster')
        if image and 'src' in image.attrs:
            # tmdb uses relative paths for posters in search
            src = image['src']
            if 't/p/' in src:
                # Ensure we get the full URL
                if src.startswith('/'):
                    return f"https://www.themoviedb.org{src}"
                return src
        
        # Fallback to general CDN search if first scrape fails
        return f"https://via.placeholder.com/300x450.png?text={requests.utils.quote(movie_title)}"
    except Exception as e:
        print(f"Error scraping {movie_title}: {e}")
        return None

def enrich_database():
    print("Connecting to database...")
    conn = sqlite3.connect(DB_PATH)
    cursor = conn.cursor()
    
    cursor.execute("SELECT id, title, release_year FROM movie WHERE image_url IS NULL OR image_url = ''")
    movies = cursor.fetchall()
    
    print(f"Found {len(movies)} movies to enrich.")
    
    for movie_id, title, year in movies:
        print(f"Enriching: {title} ({year})...")
        poster_url = get_poster_from_tmdb(title, year)
        
        if poster_url:
            cursor.execute("UPDATE movie SET image_url = ? WHERE id = ?", (poster_url, movie_id))
            conn.commit()
            print(f"  Success: {poster_url}")
        else:
            print(f"  Failed to find poster for {title}")
        
        # Rate limit to be a good citizen
        time.sleep(1)

    conn.close()
    print("Enrichment complete.")

if __name__ == "__main__":
    enrich_database()
