backend/requirements.txt
Flask
Flask-SQLAlchemy
strawberry-graphql[flask]
Faker
flask-cors
beautifulsoup4
requests
